b1.16.6 - retract libs to shared.go
